Log dropped frames at debug level, record frame statistics
The Placebo renderer drops frames if the window is not currently
visible, which leads to a huge flood of warnings in the logs that are
ultimately noise, since streaming resumes once the window is visible
again. So instead of logging a warning when a frame is dropped, it's
merely logged at verbose level to help with debugging. Additionally, we
record the number of total frames rendered and dropped, which could
be used for reporting at a later time.